Переносит функцию замены карточки/формы
