+++ src/utils/utils.js
@@ -1,5 +1,3 @@
-import Abstract from "../view/abstract.js";
-
 export const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
@@ -12,20 +10,4 @@ export const createElement = (template) => {
   return newElement.firstChild;
 };
 
-export const replace = (newChild, oldChild) => {
-  if (newChild instanceof Abstract) {
-    newChild = newChild.getElement();
-  }
 
-  if (oldChild instanceof Abstract) {
-    oldChild = oldChild.getElement();
-  }
-
-  const parent = oldChild.parentElement;
-
-  if (newChild === null || oldChild === null || parent === null) {
-    throw new Error(`Can't replace - elements don't exist`);
-  }
-
-  parent.replaceChild(newChild, oldChild);
-};
